added FileReader functionality

+++ src/Movie.java
@@ -2,7 +2,7 @@ package src;
 
 public class Movie extends Media {
     
-    public Movie(String title, int release, int rating) {
+    public Movie(String title, int release, double rating) {
         super(title, release, rating);
     }
 }
